// src/types/openapi.d.ts
export interface components {
    schemas: {
// ...
        Doc: {
            /**
             * @description ID of the Coda doc.
             * @example AbCDeFGH
             */
            id: string;
            /**
             * @description The type of this resource.
             * @enum {string}
             */
            type: "doc";
            /**
             * Format: url
             * @description API link to the Coda doc.
             * @example https://coda.io/apis/v1/docs/AbCDeFGH
             */
            href: string;
            /**
             * Format: url
             * @description Browser-friendly link to the Coda doc.
             * @example https://coda.io/d/_dAbCDeFGH
             */
            browserLink: string;
            icon?: components["schemas"]["Icon"];
            /**
             * @description Name of the doc.
             * @example Product Launch Hub
             */
            name: string;
            /**
             * Format: email
             * @description Email address of the doc owner.
             * @example user@example.com
             */
            owner: string;
            /**
             * @description Name of the doc owner.
             * @example Some User
             */
            ownerName: string;
            docSize?: components["schemas"]["DocSize"];
            sourceDoc?: Record<string, never> & components["schemas"]["DocReference"];
            /**
             * Format: date-time
             * @description Timestamp for when the doc was created.
             * @example 2018-04-11T00:18:57.946Z
             */
            createdAt: string;
            /**
             * Format: date-time
             * @description Timestamp for when the doc was last modified.
             * @example 2018-04-11T00:18:57.946Z
             */
            updatedAt: string;
            published?: components["schemas"]["DocPublished"];
            folder: components["schemas"]["FolderReference"];
            workspace: components["schemas"]["WorkspaceReference"];
            /**
             * @deprecated
             * @description ID of the Coda workspace containing this doc.
             * @example ws-1Ab234
             */
            workspaceId: string;
            /**
             * @deprecated
             * @description ID of the Coda folder containing this doc.
             * @example fl-1Ab234
             */
            folderId: string;
        };
// ...
        DocList: {
            items: components["schemas"]["Doc"][];
            /**
             * Format: url
             * @description API link to these results
             * @example https://coda.io/apis/v1/docs?limit=20
             */
            href?: string;
            nextPageToken?: components["schemas"]["nextPageToken"];
            nextPageLink?: components["schemas"]["nextPageLink"] & string;
        };
        /** @description Payload for creating a new doc. */
        DocCreate: {
            /**
             * @description Title of the new doc. Defaults to 'Untitled'.
             * @example Project Tracker
             */
            title?: string;
            /**
             * @description An optional doc ID from which to create a copy.
             * @example iJKlm_noPq
             */
            sourceDoc?: string;
            /**
             * @description The timezone to use for the newly created doc.
             * @example America/Los_Angeles
             */
            timezone?: string;
            /**
             * @description The ID of the folder within which to create this doc. Defaults to your "My docs" folder in the oldest workspace you joined; this is subject to change. You can get this ID by opening the folder in the docs list on your computer and grabbing the `folderId` query parameter.
             * @example fl-ABcdEFgHJi
             */
            folderId?: string;
            initialPage?: Record<string, never> & components["schemas"]["PageCreate"];
        };
// ...
    };
// ...
}